Introduce planex.blobs and planex.macros to simplify maintainance
